requirements.txt
requests
Pillow
python-dotenv
playwright
tkcalendar
tkinter-tooltip
